This is the database python script.
I left details in the comments.  All testing was with the outifle.json and a duplicate of it with different IDs (i.e.  I changed to 1 and 2 to 3 and 4 on the second file).  All queries generate a response.json file.  To load the database, pass it the complete json file.  if you need to append, pass another json file, but do not add duplicates and keep IDs different.  Populate flushes everything out of the database before filling it.